requirements.txt
Django>=5.1,<5.2
google-generativeai>=0.5.0
python-dotenv>=0.15.0